requirements.txt
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
Flask-Migrate==2.5.3
Flask-Script==2.0.6
flask-marshmallow==0.13.0
marshmallow-sqlalchemy==0.23.1
requests==2.24.0